Move utilities to roll/lib; render key type (b/w) by class name
