+++ piano_roll/lib/helpers.js
